+++ crates/hyperswitch_connectors/src/connectors/payload/transformers.rs
@@ -444,6 +444,7 @@ where
                         network_txn_id: None,
                         connector_response_reference_id: response.ref_number,
                         incremental_authorization_allowed: None,
+                        authentication_data: None,
                         charges: None,
                     })
                 };